+++ src/parallel_shooter/shot.go
@@ -28,8 +28,8 @@ func (s *Shot) Update() {
 		return
 	}
 	prev_y := s.point.Y()
-	s.point = NewPoint(s.point.X() + int(s.velocity.X()), s.point.Y() + int(s.velocity.Y()))
-	hitArea := NewArea(NewPoint(s.point.X(), prev_y), NewPoint(s.point.X()+s.width, s.point.Y()))
+	s.point = NewPoint(s.point.X() + s.velocity.X(), s.point.Y() + s.velocity.Y())
+	hitArea := NewArea(NewPoint(s.point.X(), prev_y), NewPoint(s.point.X()+float64(s.width), s.point.Y()))
 	for _, o := range s.enemies {
 		e, _ := o.(Common)
 		if e.Phase() == s.phase && hitArea.isHit(e.Area()) {
@@ -43,10 +43,10 @@ func (s *Shot) Draw(img *ebiten.Image) error {
 	return nil
 }
 
-func (s *Shot) X() int { return s.point.X() }
-func (s *Shot) Y() int { return s.point.Y() }
+func (s *Shot) X() int { return int(s.point.X()) }
+func (s *Shot) Y() int { return int(s.point.Y()) }
 func (s *Shot) Area() *Area {
-	return NewArea(NewPoint(s.point.X(), s.point.Y()), NewPoint(s.point.X()+s.width, s.point.Y()+s.height))
+	return NewArea(NewPoint(s.point.X(), s.point.Y()), NewPoint(s.point.X()+float64(s.width), s.point.Y()+float64(s.height)))
 }
 
 func (s *Shot) addEnemy(e interface{}) {
@@ -82,14 +82,14 @@ func (s *Shot) destroy() {
 }
 
 func (s *Shot) setCenter(a *Area) {
-	x := (a.p2.x - a.p1.x) / 2 + a.p1.x - s.width / 2
-	y := (a.p2.y - a.p1.y) / 2 + a.p1.y - s.height / 2
+	x := (a.p2.x - a.p1.x) / 2 + a.p1.x - float64(s.width) / 2
+	y := (a.p2.y - a.p1.y) / 2 + a.p1.y - float64(s.height) / 2
 	s.point = NewPoint(x,y)
 }
 
 func (s *Shot) Center() *Point {
 	a := s.Area()
-	x := (a.p2.x - a.p1.x) / 2 + a.p1.x - s.width / 2
-	y := (a.p2.y - a.p1.y) / 2 + a.p1.y - s.height / 2
+	x := (a.p2.x - a.p1.x) / 2 + a.p1.x - float64(s.width) / 2
+	y := (a.p2.y - a.p1.y) / 2 + a.p1.y - float64(s.height) / 2
 	return NewPoint(x,y)
 }
